+++ src/main/java/abd/p1/view/profile/selected/CompatibilityPanel.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package abd.p1.view.users.profile;
+package abd.p1.view.profile.selected;
 
 /**
  *
